requirements.txt
Django>=3.2.4,<5.2
djangorestframework>=3.12.4,<3.15
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3
